Add files with Comporator
