+++ DZ_14/SortByAgeAscending.java
@@ -1,0 +1,16 @@
+import java.util.Comparator;
+
+public class SortByAgeAscending implements Comparator<Student> {
+
+    @Override
+    public int compare(Student student1, Student student2) {
+
+        if (student1.getAge() < student2.getAge()) {
+            return -1;
+        }
+        if (student1.getAge() == student2.getAge()) {
+            return 0;
+        }
+        return 1;
+    }
+}
