+++ DZ_14/Student.java
@@ -9,10 +9,6 @@ public class Student {
         this.age = age;
     }
 
-    public Student() {
-
-    }
-
     public int getAge() {
         return age;
     }
